# Conflicts:
#	msa-parking-service/models/parking.py
#	msa-parking-service/routes/parking.py
#	msa-parking-service/service/parking.py

+++ msa-parking-service/service/parking.py
@@ -13,8 +13,9 @@ def create_triggers():
     with engine.connect() as connection:
         result = connection.execute(text("SELECT name FROM sqlite_master WHERE type='trigger' AND name='after_insert_parking'"))
         if not result.fetchone():
+            # 차량번호 입력 해 parking 테이블에 저장하면 trigger 작동 (parkseat에 값 저장)
             trigger_sql_parking = """
-            CREATE TRIGGER after_insert_parking
+            CREATE TRIGGER insert_parking
             AFTER INSERT ON parking
             FOR EACH ROW
             BEGIN
@@ -25,17 +26,22 @@ def create_triggers():
 
         result = connection.execute(text("SELECT name FROM sqlite_master WHERE type='trigger' AND name='remove_parkseat'"))
         if not result.fetchone():
+            # 결제완료 시 payment 테이블에 저장 된 paydate에 값이 채워지면
+            # parkseat에 동일한 carnum 데이터 삭제
             trigger_sql_payment = """
             CREATE TRIGGER remove_parkseat
-            AFTER INSERT ON payment
+            AFTER UPDATE ON payment
             FOR EACH ROW
             BEGIN
-                DELETE FROM parkseat WHERE carnum = NEW.carnum;
+                IF OLD.paydate IS NULL AND NEW.paydate IS NOT NULL THEN
+                    DELETE FROM parkseat WHERE carnum = NEW.carnum;
+                END IF;
             END;
             """
             connection.execute(text(trigger_sql_payment))
 
-
+# 입차
+# 차량번호 입력 시 parking테이블에 저장
 def register(db: Session, parking: ParkingBase):
     parking = Parking(**parking.model_dump())
     db.add(parking)
@@ -44,8 +50,8 @@ def register(db: Session, parking: ParkingBase):
 
     return parking
 
-
-def search(db: Session, parknum: str):
+# 입차 내역 전부 조회
+def carlists(db: Session, parknum: str):
     query = (
         db.query(Parkseat.carnum, Parking.intime)
         .join(Parking, Parking.carnum == Parkseat.carnum)
@@ -54,7 +60,8 @@ def search(db: Session, parknum: str):
     result = query.all()
     return result
 
-
+# 출차
+# carlists에서 주차한 차를 선택해서 outregist페이지로 넘어갈 때 outtime 저장
 def set_outtime(db: Session, carnum: str):
     parking = db.query(Parking).filter(Parking.carnum == carnum).first()
 
@@ -64,16 +71,17 @@ def set_outtime(db: Session, carnum: str):
     parking.outtime = datetime.now()
     db.commit()
 
-    total_time = parking.outtime - parking.intime
-    total_minutes = total_time.total_seconds() / 60
-
-    rate_10min = 1500       # 회차시간 15분 / 10분당 1500원
-    total_fee = int((max(0, total_minutes - 15) / 10) * rate_10min)
-
-    return {
-        "carnum": parking.carnum,
-        "intime": parking.intime,
-        "outtime": parking.outtime,
-        "total_minutes": total_minutes,
-        "total_fee": total_fee
-    }
+# 차량 정보 출력
+    # total_time = parking.outtime - parking.intime
+    # total_minutes = total_time.total_seconds() / 60
+    #
+    # rate_10min = 1500       # 회차시간 15분 / 10분당 1500원
+    # total_fee = int((max(0, total_minutes - 15) / 10) * rate_10min)
+    #
+    # return {
+    #     "carnum": parking.carnum,
+    #     "intime": parking.intime,
+    #     "outtime": parking.outtime,
+    #     "total_minutes": total_minutes,
+    #     "total_fee": total_fee
+    # }
